Объединяет директорию utils в единый модуль
