--- js/utils/check-max-length.js
@@ -1,7 +1,0 @@
-/**
- * Проверяет соответствие длины строки максимально допустимому значению.
- * @param {string} string - Проверяемая строка.
- * @param {number} maxLength - Максимально допустимое значение длины строки.
- * @returns {boolean} Возвращает true, если длина строки не превышает максимально допустимого значения. Возвращает false, в остальных случаях.
- */
-export const checkMaxLength = (string, maxLength) => string.length <= maxLength;
