--- js/utils/get-random-integer.js
@@ -1,15 +1,0 @@
-/**
- * Возвращает случайное целое число в заданном диапазоне.
- * При передаче дробных значений границ диапазона учитывается только целая часть.
- * @param {number} min - Минимальное значение диапазона.
- * @param {number} max - Максимальное значение диапазона.
- * @returns {number} Результат всегда больше или равен минимальному значению диапазона.
- */
-export const getRandomInteger = (min, max) => {
-  max = Math.trunc(max);
-  min = Math.trunc(min);
-
-  const range = max > min ? max - min + 1 : 1;
-
-  return min + Math.floor(range * Math.random());
-};
