--- js/utils/get-shuffled-integer-sequence.js
@@ -1,35 +1,0 @@
-/**
- * Перемешивает элементы массива случайным образом.
- * Использует алгоритм тасования Фишера-Йетса (источник: https://learn.javascript.ru/task/shuffle)
- * @param {Array} array - Исходный массив.
- * @returns {Array} Вовзращает перемешанную неглубокую копию исходного массива.
- */
-export const shuffle = (array) => {
-  array = [...array];
-
-  for (let i = array.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-
-    [array[i], array[j]] = [array[j], array[i]];
-  }
-
-  return array;
-};
-
-/**
- * Возвращает упорядоченный массив целых чисел из заданного диапазона.
- * При передаче дробных значений границ диапазона учитывается только целая часть.
- * @param {number} min - Минимальное значение диапазона.
- * @param {number} max - Максимальное значение диапазона.
- * @returns {Array} Массив состоит как минимум из одного элемента равного минимальному значению диапазона.
- */
-export const getIntegerSequence = (min, max) => {
-  max = Math.trunc(max);
-  min = Math.trunc(min);
-
-  const range = max > min ? max - min + 1 : 1;
-
-  return new Array(range).fill(min).map((item, index) => item + index);
-};
-
-export const getShuffledIntegerSequence = (min, max) => shuffle(getIntegerSequence(min, max));
